refactor: improve responsive design
The padding of the main container and footer were reduced to 1rem. The main container now has a gap of 1rem between its children when the screen is wider than 1200px. The header and footer now have a padding of 1rem on screens smaller than 992px, and a padding of 1rem on the sides for screens larger than that. Additionally, the header's flex-direction changes to row on screens larger than or equal to 768px.